Added ConVar tf_weapon_forced_critical_distance_falloff
This ConVar enables forced weapon critical distance falloff. The Ambassador is the only weapon currently affected by this, but it is just a tag in items_games.txt. This allows us to add it to other weapons while giving server operators options to change it, which will give us more flexibility in the long run.