Tick aligned raw mouse input.
Adds an improved version of m_rawinput. This is enabled by setting it to 2. Frame rate instabilities do not affect mouse input sampling in this mode.
Part of the jank is because it was made through reverse engineering a few functions. This is still rather experimental and may need some improvements in the long run.
I will comeback soon to improve the structure of the modifications.